fix: make result_t pointer-sized to fix gitclient crash on arm64
result_t was uint32_t (32-bit), which truncated pointers on arm64
when tvGetSelectedRecord returned a record pointer through send_message.
Changed result_t to intptr_t and send_message to return intptr_t.

Also disabled gitclient debug logging by default (GITCLIENT_DEBUG=0).

+++ tests/test_env.c
@@ -147,7 +147,7 @@ window_t* test_env_create_window(const char *title, int x, int y, int w, int h,
 }
 
 // Helper: Send a tracked message
-int test_env_send_message(window_t *win, uint32_t msg, uint32_t wparam, void *lparam) {
+intptr_t test_env_send_message(window_t *win, uint32_t msg, uint32_t wparam, void *lparam) {
     return send_message(win, msg, wparam, lparam);
 }
