+++ tests/test_env.h
@@ -65,7 +65,7 @@ window_t* test_env_create_window(const char *title, int x, int y, int w, int h,
                                   winproc_t proc, void *userdata);
 
 // Helper: Send a tracked message
-int test_env_send_message(window_t *win, uint32_t msg, uint32_t wparam, void *lparam);
+intptr_t test_env_send_message(window_t *win, uint32_t msg, uint32_t wparam, void *lparam);
 
 // Helper: Post a tracked message
 void test_env_post_message(window_t *win, uint32_t msg, uint32_t wparam, void *lparam);
